tpl/data: Move cache logic to separate file
